Added separate load_config for the API
